Check Permission Codes when Rendering Admin Link
