+++ src/herndon_law_seo_helper/herndon_law_seo_helper/production_settings.py
@@ -68,7 +68,8 @@ TEMPLATES = [
                 'django.template.context_processors.request',
                 'django.contrib.auth.context_processors.auth',
                 'django.contrib.messages.context_processors.messages',
-                'scraper_manager.context_processors.profile_image.get_profile_image'
+                'scraper_manager.context_processors.user_context.get_user_permissions',
+                'scraper_manager.context_processors.user_context.get_profile_image',
             ],
         },
     },
