+++ src/herndon_law_seo_helper/scraper_manager/templates/base.html
@@ -33,9 +33,11 @@
               <li class="nav-item">
                 <a class="nav-link text-white" href={% url 'dashboard' %}>Dashboard</a>
               </li>
-              <li class="nav-item">
-                <a class="nav-link text-white" href={% url 'admin_dashboard' %}>Admin</a>
-              </li>
+              {% if can_view_admin %}
+                <li class="nav-item">
+                  <a class="nav-link text-white" href={% url 'admin_dashboard' %}>Admin</a>
+                </li>
+              {% endif %}
               <li class="nav-item">
                 <a class="nav-link text-white" href={% url 'settings' %}>Settings</a>
               </li>
